Delete VPC resources only if they are managed by Kore (tag: kore.appvia.io/managed=true)

+++ pkg/utils/cloud/aws/route_table.go
@@ -140,6 +140,10 @@ func DeleteRouteTable(svc ec2.EC2, vpc VPC, rtType, az string) error {
 		return nil
 	}
 
+	if !IsKoreManaged(routeTable.Tags) {
+		return nil
+	}
+
 	_, err = svc.DeleteRouteTable(&ec2.DeleteRouteTableInput{
 		RouteTableId: routeTable.RouteTableId,
 	})
